requirements.txt
absl-py==2.1.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.8.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blessed==1.20.0
blinker==1.8.2
boto3==1.35.21
botocore==1.35.21
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deepface==0.0.93
Django==5.1.1
django-picklefield==3.2
django-storages==1.14.4
djangorestframework==3.15.2
filelock==3.16.1
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.45.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
h5py==3.11.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jinxed==1.3.0
jmespath==1.0.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==0.1.1
namex==0.0.8
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==4.25.4
psycopg2-binary==2.9.9
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==3.5.3
requests==2.32.3
retina-face==0.0.17
rich==13.8.1
s3transfer==0.10.2
setuptools==75.1.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
tf_keras==2.17.0
tqdm==4.66.5
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
whitenoise==6.7.0
wrapt==1.16.0
